requirements.txt
discord.py
easy-pil
typing-extensions
